added sent email feature

+++ server/src/controllers/email.ts
@@ -1,18 +1,19 @@
 
 import Email from '../models/email';
-
+import user from '../models/user';
 export const sendEmail = async (req:any,res:any) =>{
-   
+    const reciever =await user.findOne({email:req.body.reciever}); 
+    console.log(reciever);
     const newEmail = new Email({...req.body,sender:req.user.email,
-        senderId:req.user._id, senderName:req.user.name});
+        senderId:req.user._id.toString(), senderName:req.user.name , recieverName: reciever.name});
     await newEmail.save();
-    res.json({message: 'Email sent successfully'});
+    res.json(newEmail);
 
 }
 export const inbox = async (req:any,res:any) =>{
     
     try {
-        console.log(req.user.email);
+        // console.log(req.user.email);
         const inbox = await Email.find({reciever:req.user.email});
         // console.log(inbox);
         res.json(inbox);
@@ -32,12 +33,39 @@ export const markEmailAsRead = async (req:any , res:any) =>{
         console.log(err);
     }
 }
-export const deleteEmail = async(req:any,res:any) =>{
-    const emailId = req.params.id;
-    console.log(emailId);
+export const getSendEmails = async (req:any , res:any) =>{
     try{
-        await Email.findByIdAndDelete({_id:emailId});
-        console.log('email deleted');
+        const sentEmails = await Email.find({senderId:req.user._id});
+        // console.log(sentEmails);
+        res.json(sentEmails)
+        // console.log(req.user);
+    }
+    catch(err){
+        console.log(err)
+    }
+}
+export const deleteRecievedEmail = async(req:any,res:any) =>{
+    const emailId = req.params.id;
+    // console.log(emailId);
+    try{
+        const email = await Email.findById({_id:emailId});
+        if(email.sender==='') await Email.findByIdAndDelete({_id:emailId});
+        else await Email.findByIdAndUpdate({_id:emailId},{reciever:""});
+        // console.log('email deleted');
+        res.json({message:'Email deleted successfully'});
+    }
+    catch(err){
+        console.log(err);
+    }
+}
+export const deleteSentEmail = async(req:any,res:any) =>{
+    const emailId = req.params.id;
+    // console.log(emailId);
+    try{
+        const email = await Email.findById({_id:emailId});
+        if(email.reciever==='') await Email.findByIdAndDelete({_id:emailId});
+        else await Email.findByIdAndUpdate({_id:emailId},{sender:"", senderId:"NA"});
+        // console.log('email deleted');
         res.json({message:'Email deleted successfully'});
     }
     catch(err){
